extract removeRedundant to an external file
